requirements.txt
asgiref==3.3.1
attrs==20.3.0
Django==3.1.4
django-ckeditor==6.0.0
django-crispy-forms==1.11.0
django-js-asset==1.2.2
django-multiselectfield==0.1.12
django-schema-graph==1.2.0
django-widget-tweaks==1.4.8
gunicorn==20.1.0
Pillow==8.1.0
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1